Fix broken links in the doc index page
In the `/v1.15` page, we have a section called `Use Bundler with`,
where we have some links that was intended to point to the guides.
But unfortunately those links are broken and still seems to point
to the old html file that does not exist anymore.

This commit fixes that and point those broken link to the correct
guide path that changed in PR #322.