Minor updates on the deploying guide
In the deploying guides, `/guides/deploying.html` there seems to
be a broken link, the `executables` does not seems to be linking
to the correct path. And, the guide also using a sample from the
old Heroku's way of adding configuration.

This commit fixes that link to the correct path, and it also
updates the heroku sample to their latest way of configuring
environment variable.

Ref: https://devcenter.heroku.com/articles/bundler