Clarify and deepen the discussion of Gemfile.lock in a gem repository
Based on a discussion on GitHub:
https://github.com/bundler/bundler/issues/5879#issuecomment-551967102

This patch updates the documentation from saying "do this one thing" to explaining the pros and cons of committing a Gemfile.lock to source control.  This is important -- not only in light of the fact that the conventional wisdom has reversed itself in the past few years, but because what works for open source gems may not be best for privately-maintained gems.